+++ test/open_pgp/packet_test.exs
@@ -1,4 +1,5 @@
 defmodule OpenPGP.PacketTest do
   use OpenPGP.Test.Case, async: true
   doctest OpenPGP.Packet
+  doctest OpenPGP.Encode.impl_for!(%OpenPGP.Packet{})
 end