+++ test/open_pgp/util/pkcs1_test.exs
@@ -1,0 +1,4 @@
+defmodule OpenPGP.Util.PKCS1Test do
+  use OpenPGP.Test.Case, async: true
+  doctest OpenPGP.Util.PKCS1
+end